2013-01-28  Patrick Eriksson  <patrick.eriksson@chalmers.se>
            * arts-2-1-469

            * controlfiles/general/agendas.arts: Adopted definition of
            ppath_agenda__TransmitterReceiverPath.

            * src/m_ppath.cc (ppathFromRtePos2): Now switched to use new
              version
              of the method. (Old version deleted)

              * src/m_transmitter.cc (iyRadioLink): Now also accepting
                "unvalid"
                ppaths. This gives iy=NaN.

                * src/ppath.cc (ppath_set_background,ppath_what_background):
                Defined string for case 0 to be "unvalid" (before just "").

                * src/lin_alg.cc: Added type casting where dividing with an
                  integer
                  number.



git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@8147 aaf1aab0-4228-0410-ad68-8dceda47f409

+++ src/lin_alg.cc
@@ -416,7 +416,7 @@ void linreg(
   Numeric s1=0, xm=0, s3=0, s4=0;
 
   for( Index i=0; i<n; i++ )
-    { xm += x[i]/n; }
+    { xm += x[i]/Numeric(n); }
 
   for( Index i=0; i<n; i++ )
     {
@@ -427,5 +427,5 @@ void linreg(
     }
     
   p[1] = s1 / s4; 
-  p[0] = s3/n - p[1]*xm;
+  p[0] = s3/Numeric(n) - p[1]*xm;
 }
